// src/interfaces/appInterfaces.tsx
export interface Data {
    user: User;
}

export interface User {
    id: number;
    dni: string;
    name: string;
    phone: string;
    password: string;
    roles: Role[];
}

export interface Role {
    id: number;
    name: string;
    guard_name: string;
    created_at: string;
    updated_at: string;
    pivot: Pivot;
}

export interface Pivot {
    model_id: number;
    role_id: number;
    model_type: string;
}

export interface Msg {
    summary: string;
    detail: string;
    code: string;
}

// Generated by https://quicktype.io

export interface LoginData {
    dni: string;
    password: string;
}


// Generated by https://quicktype.io

export interface RegisterResponse {
    data: Data;
    token: string;
    msg: Msg;
}
// ...
export interface UsersResponse {
    data: Usuarios[];
    summary: string;
    detail: string;
    code: string;
}

export interface Usuarios {
    id: string;
    created_at?: string;
    updated_at?: string;
    dni: string;
    name: string;
    phone: string;
}

// Generated by https://quicktype.io

export interface UsuarioResponse {
    data: Data;
    msg: Msg;
}